+++ CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Changed
+
+- services/network; Camel-Cased HTTP headers
+
+### Fixed
+
+- services/network; made HTTP header comparison case-insensitive
+
 ## [1.3.0]
 
 ### Changed